GUACAMOLE-300: New function UserService::getUidFromUserDN
Determine the username which corresponds to the user having the given DN.
Assume that a userDN record contains at least one attribute identifying a
username, and that if it contains more than one such attribute then it's
the SAME username.
Also assume that each attribute contains single value, not array.